turn osp_time_{ms,secs} into 64 bit return functions

+++ MusicFan/MusicFan/MFANAqPlayer.mm
@@ -175,8 +175,8 @@ MFANAqPlayer_getUnknownString()
      * doesn't notice if you actually pause playing (another Apple
      * POS).
      */
-    uint32_t _lastReturnedMs;	/* time an audio packet was last returned to us */
-    uint32_t _lastDataMs;	/* last time data was received */
+    uint64_t _lastReturnedMs;	/* time an audio packet was last returned to us */
+    uint64_t _lastDataMs;	/* last time data was received */
     uint32_t _lastDataBytes;	/* bytes since lastDataMs */
     uint32_t _formatId;		/* audio format ID */
     float _dataRate;		/* estimated data rate */
@@ -399,7 +399,7 @@ MFANAqPlayer_handleOutput( void *acontextp,
  */
 - (void) spyMonitor: (id) junk
 {
-    uint32_t now;
+    uint64_t now;
 
     pthread_mutex_lock(&_aqMutex);
 
@@ -890,7 +890,7 @@ MFANAqPlayer_rsDataProc(void *contextp, RadioStream *radiop, char *bufferp, int3
     OSStatus osStatus;
     std::string *contentTypep;
     AudioFileTypeID fileType;
-    uint32_t now;
+    uint64_t now;
     float updateRate;
     uint32_t delta;
